src/requirements.txt
Django==2.1.1
django-cors-headers==2.4.0
django-debug-toolbar==1.9.1
djangorestframework==3.8.2
hashids==1.2.0
psycopg2==2.7.5